resourceHandlingAtom moved to separate file
